// src/image-upload/dto/aws-file.dto.ts
export interface AWSFile {
  fieldname: string;
  originalname: string;
  encoding: string;
  mimetype: string;
  ACL: string;
  ETag: string;
  Location: string;
  key: string;
  Key: string;
  Bucket: string;
  width: number;
  height: number;
  premultiplied: boolean;
  size: number;
  ContentType: string;
}